refactor: Remove deprecated function in test-utils (#14229)
## Description

Removes the deprecated `ensureContainerConnected` function from
`@fluidframework/test-utils` and changes the default value of the
`failOnContainerClose` parameter of `waitForContainerConnection` from
`false` to `true`.

## Breaking Changes

Yes. We announced these changes [in release
2.0.0-internal.2.4.0](https://github.com/microsoft/FluidFramework/blob/main/BREAKING.md#deprecate-ensurecontainerconnected-in-fluidframeworktest-utils).
It was too close to 2.0.0-internal.3.0.0 to actually make the breaking
change there, so it's going in now, to be included in
2.0.0-internal.4.0.0.

## Reviewer guidance

A bunch of the changes in this PR are just cleaning up existing calls to
`waitForContainerConnection()` that don't need to pass `true` explicitly
as second parameter.

I wasn't sure if this should be a `refactor`, or a `feat` for commitlint
purposes. Feedback welcome.