Make summarizer client read-only to layers below container runtime (#14354)
Summarizer client should never have local changes as it can lead to
eventual consistency and potential other issues. In order to prevent
user / app changes in the summarizer client that can lead to this, the
summarizer client should be in read-only mode.

This change adds a delta manager proxy at the container runtime layer
and is passed to layers below. The delta manager proxy returns read-only
for summarizer clients. This will help users / app prevent making local
changes in summarizer by checking read-only state.

Note that summarizer has to send summary op and upload the summary but
those are are the container runtime layer which uses the actual delta
manager and not the proxy.