Render Container Data in Client Debugger in Tree Format  (#14380)
#### Description
https://dev.azure.com/fluidframework/internal/_workitems/edit/2929/
https://dev.azure.com/fluidframework/internal/_workitems/edit/3534/

- Provide tree-view of Client Debugger's container data debug panel
- Provide data type of data inside the panel 

This PR serves as a prototype for creating a custom renderer for
Container Data in the future. Currently, we are taking a look at
leveraging `summarize` for retrieving a JavaScript object which contains
entire data inside the single container. If this is viable, we will be
creating a custom tree-view renderer which will be able to traverse
through Fluid objects and propagate necessary changes.

#### Before 


#### After 

![image](https://user-images.githubusercontent.com/111468570/222293735-b63c5837-8050-451b-bf99-56f4c8e37bad.png)