Deprecate Interval APIs intended to be internal (#14318)
## Description

inheritdoc doesn't copy release tags. This marks functions implementing
an internal method in `Interval` and `SequenceInterval` deprecated, so
we can mark them internal in a future release.

Ideally we would clean up API surface here; exporting the classes isn't
great in the first place and there's not a strong reason we need to.