Reenable chunking in the stress tests (#14417)
## Description

There have been some attempts to enable chunking in the stress tests,
however they were reverted due to various reasons, such as:
- they were creating too many chunked ops overwhelming the other clients
and no client could catch up
- they were growing both the file size and the snapshot size to
unreasonable values

This PR would enable chunking but will keep traffic low enough to not
affect regular testing but high enough to exercise chunking. After
https://github.com/microsoft/FluidFramework/pull/14283, we can also keep
both the file size and the snapshot size within reasonable limits.

## How this works

Only the first 12 runIds would send ops which end up compressed and
chunked. Per the current config, the estimated total traffic would be
(equal to the size of the file): `12 x 6 x 214800 = 15.4 MB`. As we
overwrite the key in the map for which we create the large ops, the
snapshot size increases only by 214800 bytes, which is the size of the
large op.