Add Transaction APIs to ISharedTreeCheckout (#14442)
[How contribute to this
repo](https://github.com/microsoft/FluidFramework/blob/main/CONTRIBUTING.md).

[Guidelines for Pull
Requests](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).

## Description

This adds some new APIs to `ISharedTreeCheckout` which allow for the
creation of transactions. Transactions allow edits to be grouped into
atomic units that get squashed into a single commit when the transaction
completes. This required adding some new functionality to `EditManager`,
teasing apart the `SharedTreeCore.submitEdit()` function into two
distinct parts, and adding a small helper class for managing stacks of
transactions (because transactions can "nest").

* Moves `SharedTreeBranch` into shared-tree-core. Although it could
possibly be moved to its own library later, it shares a lot of patterns
with `SharedTreeCore` and simplifies the dependency graph to keep it
there for now.
* Adds a TEditor generic parameter to `SharedTreeCore` (and
`SharedTreeBranch`), but removes the `TChangeFamily` parameter to avoid
feeling guilty. `TChangeFamily` is not necessary as the `changeFamily`
field is now private.

For simplicity and to reduce noise, this PR does not attempt to remove
the old style of transaction (`ITransactionCheckout`); that will be done
in a follow-up PR.