Add a new API (empty implementation) in Alfred for token revocation (#14363)
This PR add a new API in Alfred for token revocation feature. Currently
it is an empty implementation and only return 503.
It also adds a new scope (permission) for this new API: token:revoke
This new scope is not in the same place as other scopes because scopes
are usually defined in protocol package, which is out of server package.
Not sure if it should be moved there. Or we can move it later in the
future.

The API url is like this:
${alfred_url}/documents/${tenantId}/document/${documentId}/revokeToken

This API can be invoked only if the token include the new scope:
token:revoke.

---------

Co-authored-by: Yunpeng Hou <yunho-microsoft@Yunpengs-MacBook-Pro-2.local>
Co-authored-by: Yunpeng Hou <yunho-microsoft@Yunpengs-MBP-2.guest.corp.microsoft.com>