[GC-stress] Added snapshot cache expiry override to ODSP driver (#14454)
Snapshot cache expiry is current disabled in snapshot tests. We need
snapshot cache to test scenarios where a summarizer loads from a cached
snapshot and has to update its summary / GC state from a later summary
ack. We have recently seen bugs because of these scenarios and the
stress tests should run these.

Added an option to override the snapshot cache expiry timeout in ODSP
driver. Updated the test config to use that option. The snapshot cache
expiry is configured to be half of max faultInjectionMs. The idea is
that on average half the sessions that fail because of faultInjection
will load from cached snapshot and other half will not.


[AB#3315](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3315)