Fix stress test op count (#14499)
## Description

ADO:3647

When the stress test clients receive a nack, the client will reconnect
and the number of ops it has already sent will reset. This means that
for every nack and depending on when it actually happens (within the op
send window), the client will send more ops than configured, as it
thinks it starts fresh.

In the worst case scenario, the test is configured to send 10k ops with
120 clients => each client sends ~80 ops and each client receives a nack
at around sending op number 79. The test will end up sending ~20k ops,
which will double for every time this happens to all clients.

Additionally, if one of the clients keeps getting nacked, it will send
ops until it reaches 80 without getting nacked. This means it's possible
(but improbable) for a client to run forever.