[main > release/v2int/3.0]: Don't proceed to close socket if connection is already disposed in odsp (#14489)
## Description

ADO item link:
https://dev.azure.com/fluidframework/internal/_workitems/edit/3648
1.) Adding more telemetry in flows where the socket is getting closed on
connect/reconnect to better understand the state of the system.
2.) We see that connect_error or some other error is fired on socket and
we go through the socket close flow and telemetry shows that we do
successfully close the socket and remove listeners but then we again
have the errror on socket which again go into the socket close flow and
cause reentrancy issues with uncaught 0x416 assert. Seems like there is
a possibility that there was some micro task already scheduled by socket
to call the error listener which causes us to go through the socket
close flow again. In the end, the initial call to recoonect in
connectionManager.ts after disconnect hangs and we never try to connect
again.