Remove SharedTree legacy transaction APIs (#14469)
[How contribute to this
repo](https://github.com/microsoft/FluidFramework/blob/main/CONTRIBUTING.md).

[Guidelines for Pull
Requests](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).

## Description

This updates the transaction APIs to be more cohesive. It replaces the
use of the old `Checkout` (or `TransactionCheckout`, as it was aliased
in practice) with the new transaction API on `ISharedTreeCheckout`. It
provides a convenience helper `runSynchronous` to help with the common
case of a simple, short-lived transaction. It moves
`runSynchronousTransaction` (renamed to `runTransactionOnForest` to be a
test utility, as it is still useful for `TestTree` and forest unit
tests.

## Breaking Changes

* This removes `runTransaction` on ISharedTreeCheckout. The equivalent
replacement is the free function `runSynchronous`.