Adding error sanitizer for Mongo. (#14447)
## Description
When Insert request fails Mongo returns the data to be inserted back in
the error response. The data might contain customer content and should
be removed from error response for privacy compliance reasons. The error
sanitizer uses the fast-redact package to redact the keys at specified
paths.

Tested in a dev cluster, works as expected.