[GC-stress] Use container options instead of configs for most GC options (#14514)
Updated the tests to use container runtime GC options instead of
configurations. The advantage of using gc options over configs is that
most of them are persisted in the summary and won't change for the
clients. For example, configs like gcEnabled, sweepEnabled, session
expiry timeout and session timeout should not change for a file or else,
it will lead to errors.

Also, reverted some of the changes that removed randomization of configs
and container options. Since we are using container options override
now, randomization will be fine.