build(deps): Update ts-morph (#14526)
## Description

Update ts-morph. Version 13 is needed to cover TypeScript 4.5 used in
the repo, but this updates to the newest version, 17.0.1 to future proof
it for future typescript upgrades.

Also bumps version in build-cli readmes which were missed in last
version bump.

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).