ci: Remove commitlint PR validation (#14574)
This change removes the CI job that verifies commit titles match our
conventional commits config.

We have piloted conventional commits in the build-tools release group,
and the major drawback we’ve found is that the commit message, once
written, cannot be changed. This means that when we do a release and
generate the changelog, any tweaks we want to make have to be made and
stored elsewhere. Ideally, we want to be able to review and fix change
announcements just like code.

The other major drawback is that because the matadata is in the commit
message itself and cannot be changed, we need to have some CI validation
to ensure that the commit messages are in the right format. Otherwise
they become “invisible” to tools reading the log. Verifying commit
messages in GitHub CI is inefficient.

The config is still present in build-tools, and we may move it to the
root so it's easier for people to use if they want to.