v2int 3.3 Port: Improve safety when using and combing app and protocol tree (#14548) (#14560)
## Description
There is a bug in the code when a detached container with blobs is
uploaded with the single commit summary flow. The bug results from
combineAppAndProtocolSummary being called twice on the same summary.
This change fixes that issue by improving typing and defensiveness when
using and combing app and protocol trees, such that double wrapping can
be detected, and thrown as an error. The fix for the specific issue is
in

`packages\loader\container-loader\src\protocolTreeDocumentStorageService.ts`.
Additionally, i've updated a number of usage of the combined tree to
remove unsafe casting.

Validated the issue via test modifications, which now pass with fixes. 
Fixes
[AB#3711](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3711)
Related
[AB#3717](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3717)