Small adjustments to the E2E infra to support another set of documents (multiple DDSs) (#14528)
## Description

There was a recent regression during the summarization in which the UI
would hang while running the GC code path. Navin came up with a good
test case to identify the regression and shared it with us. This PR
prepares the infra to support a new test case for the E2E tests. Some
changes were required to allow us to have the same existing tests (Load
and Summarization) loading all the pre-defined set of documents: Ex. Map
and Multiple DDSs as they required a different number of min iterations.

@agarwal-navin - I'll release another version with the new test to add
the Multiple DDSs.