build: Upgrade pipelines to Node16 + upgrade `nvm` recommendation (#14304)
## Description

Refreshing https://github.com/microsoft/FluidFramework/pull/13807

Update the recommended versions of Node.js for developers to use, as
well as the version used by CI.

This updates from version 14 (which goes out of support on 2023-04-30)
to version 16 (which goes out of support on 2023-09-11).

The minimum, as enforced by the root `package.json`'s `"engines"` block
is not updated to 16 yet. This can be done after a migration period.

This has no impact on the produced packages, but will allow (but not yet
require) developers working in this repo to upgrade.

Users of
[nvm](https://npm.github.io/installation-setup-docs/installing/using-a-node-version-manager.html)
can upgrade by running: `nvm use && rm -r ./node_modules/ && pnpm i` in
the root of the repo. Deleating the node_modules directory is needed to
trigger reinstalling of nodegit to avoid an "Error: Module did not
self-register" when building our codebase.

Some of the packages still use `npm` (and not `pnpm`). If working on
those packages, you may want to downgrade `npm` to version 6 after
upgrading node so that you avoid converting package-lock files to the
new format unexpectedly.

## Reviewer Guidance

Feedback on this process, as well as the actual implementations / diff
is encouraged.

Questions:
1. Do we want to wait for the rest of the usages of npm to migrate to
pnpm? (Perhaps before this, or maybe just before updating the minimum
version?)
2. Is there anywhere in our docs that needs updating?

---------

Co-authored-by: Craig Macomber <Craig.Macomber@microsoft.com>