Detached container with blobs using local driver (#14291)
Current implementation of attaching a detached container with blobs is
only possible in ODSP. We need to enable it in local driver as well to
improve our test capacity. To support this, createContainer method in
localDocumentServiceFactory.ts has to run once we get a summary.
For local driver implementation, this design delays the createContainer
method by creating a new class with it. It creates a new instance of
this class while serviceFactory.createContainer since
localDocumentServiceFactory has all required objects for running the
method but it is actually run after summary is created in detach ->
attach with blobs flow in container.ts during uploadSummaryWithContext.
I chose this design having as goal to change as less code as possible in
container.ts during attach() (100% achieved) since from a driver point
of view we should be able to enable this feature by only changing
driver/server code.
## Reviewer guide 
1.(solved) I know this design has potentially concurrency issues given
the fact that I'm relying on a new static class. I would like to have
feedback on it given that locally all blobs.spec.ts tests pass. Should
we have a UT which tries to attach multiple containers at the same time
making this design to (probably) fail? Are we okay with our current
tests?
2.(solved) Tests fail because we need to make server changes first and a
server release. Besides I need to fix which drivers run
detached->attached tests. Not moving forward until I have positive
feedback on the design.