resolve the installation dependencies error (#14572)
Currently the Test Stability pipeline has encountered a failure during
the `Install Dependencies` step on both the main and next branches. The
issue can be resolved by temporarily disabling the cache pnpm store (the
[result](https://dev.azure.com/fluidframework/internal/_build/results?buildId=139828&view=results)
on my own test branch). Considering the cache is intended to save time
but the Test Stability pipeline is scheduled to run only on weekends. It
is feasible to remove the cache temporarily from the Test Stability
pipeline (while keeping it in the Build Client pipeline) to allow the
pipeline to operate normally.