Use custom benchmarking code instead of Benchmark.js (#14597)
## Description

Benchmark.js does a lot of complicated compatibility workarounds, no
longer relevant performance tweaks, tricks for timing that have better
replacements now, a lot of stuff we don't need (suites, fancy eventing
etc) as well as a bunch of stuff I just couldn't figure out (lots of
object cloning, and strange control flow). After attempting to simplify
benchmark.js and port it to typescript, eventually I ended up replacing
every piece of it.

The main benefit of this is it will be much easier to maintain and
extend. No one was actively maintaining Benchmark.js, and even if
someone was, the changes we would want would be unlikely to be accepted
due to it having different goals from our needs and changes would be
hard. The new code is much simpler and easy for us to change.

It should also be practical to make this work with webpack and browsers
now. Full support for this is not included in this change, but could be
added in the future. What's missing is mainly an ESM build and tests.

## Breaking Changes

onCycle (which is almost never used well) was incorrectly marked as
async before this change. I'm pretty sure it wasn't properly awaited,
and now is typed as synchronous and renamed to `onBatch`.

Some of the exported types have also been refactored a bit.