Fix for assert 0x244 in odsp driver (#14619)
## Description

This was a potential bug where before cleaning everything in the delta
connection object we let the user of object know about the
disconnection. Looking into the recently added telemetry, the assert
occurs after emitting disconnect on the connection object and I see that
the disconnectCore login on OdspDocumentDeltaConnection does not get
executed and the socket reference/socket removeReference code does not
execute, so later on reconnection there is a potential that we get the
same socket reference and see the socket as connected because the
pendingInitialConnection does not even get set to false.
The emission of disconnect was after cleanup but seems like with this
PR: https://github.com/microsoft/FluidFramework/pull/12223 it was moved
before that and could have caused this bug.