Optimize Gitrest latest summary pre-compute (#14535)
## Description

We want to pre-compute latest summary after every time that we write a
summary in Gitrest. However, reading the latest after writing it is
inefficient when we tend to have most, if not all, the necessary
information during the write process, so we can compute and send back
the latest full summary to Historian without reading it from storage.