[GC-stress] Report if test failed because of GC failures or not (#14649)
There are some test stress test runs that are failing but I don't see
any of the GC error / event logs in the console or in Kusto. I believe
the test is failing due to some other issue but its hard to tell the
difference.

This PR changes the return code in case of GC failures to -2. The test
result builder uses this value to alter the failure message depending on
whether the failure is because of GC errors or not.