fix(container-runtime): Nullish coalescence bug introduced by eslint (#14593)
This bug was introduced in #11977 via eslint. The eslint rule in
question is
[prefer-nullish-coalescing](https://github.com/typescript-eslint/typescript-eslint/blob/main/packages/eslint-plugin/docs/rules/prefer-nullish-coalescing.md).
In this case `<boolean | undefined> || <boolean>` was updated to use
null-ish coalescing, even though it changes the behavior (`false ||
true` => `true`, while `false ?? true` => `false`).

Supposedly, this auto-fix bug was already resolved in
`typescript-eslint` (see [this
issue](https://github.com/typescript-eslint/typescript-eslint/issues/1768),
but the behavior seems to have regressed. I will be separately
attempting to update our plugin dependencies to see if doing so fixes
this, otherwise I will file a new bug for the plugin so we don't stumble
into this in the future.