feat(check:policy): Add setting to ignore single-package pnpm workspaces (#14656)
pnpm does not support nesting packages under a workspace that is not
managed by the workspace. That is, if there's a `pnpm-workspace.yaml`
file anywhere in the parent hierarchy, pnpm doesn't install the package
individually like one might expect.

Because we have the client release group at the root of the repo,
there's a `pnpm-workspace.yaml` file in the hierarchy for our
independent packages as well. We put a workspace file in each
independent package so pnpm treats the project as a one-package
workspace.

However, for policy-check, we need to be able to treat these
single-package workspaces differently. In particular we don't want to
enforce the preinstall script that we use in release group roots. This
PR adds a new setting to policy-check that contains a list of packages
that are single-package workspaces.