feat(build-cli): Add exec command (#14635)
This PR adds an `exec` command to build-cli, which can be used to
execute shell commands in the context of packages or release groups in
the repo.

The following example runs `pnpm -r remove @rushstack/eslint-config` on
all independent packages and all release group root packages:

```shell
flub exec --all --releaseGroupRoots "pnpm -r remove @rushstack/eslint-config"
```

I think that for this command to be useful, one needs to be able to
scope the packages that it runs on to the following groups:

1. All independent packages in the repo - use the `--packages` flag.
1. All packages in the repo using a particular scope - use the `--scope`
and `--skipScope` flags.
1. A single package in a directory - use the `--dir` flag.
1. All individual packages in a release group - use the `--releaseGroup`
flag.
1. All release group root packages - use the `--releaseGroup` and
`--releaseGroupRoots` flags together.
1. All release group root packages and independent packages - use the
`--all` and `--releaseGroupRoots` flags together.
1. All release group packages and independent packages - use the `--all`
flag.

I implemented this by adding two new flags to the BasePackageCommand
class, `--all` and `--releaseGroupRoots`. Most of the code change was in
the base class. This means that the `typetest` command can also benefit
from the new flags.