Fix eventual consistency issues in SharedDirectory (#14226)
## Description

**1.)** Add a list of clientids which have created this directory. Add a
directory seq number which is set to sequence number of
create op of that directory or 0 in case it is loaded from snapshot. Add
a **isMessageForCurrentInstanceOfSubDirectory** which only process ops
if the remote message is either from the creator of directory or this
directory was created when container was detached or in case this
directory is already live(known to other clients) and the op was created
after the directory was created.
It basically tells whether a given op was created for this current
instance of subdirectory. If not, which means this op was created before
it, then we just want to drop that op on the floor. Also added this info
in the summary, so that if the new client loads, then it should have
info about the creators clientids and create sequence number, so that it
can correctly tell whether new messages are for current instance of sub
directory or not.

**2.)** If a delete sub directory op comes and there is already local
pending create/delete subdirectory, then instead of ignoring the
delete op, clear the non-pending keys which were already known to
clients. Also reset the creator clientids and seq number to 0 as the new
create op will reset them and we will also not process any
set/clear/delete ops before the new create op. clearExceptPendingKeys()
also had a bug where if a key was deleted and it was in pendingKeys()
too due to a local clear, then we were wrongly setting undefined as
value on that key.

**3.)** If there is pending delete op for this subDirectory, then don't
apply the storage ops as we are going to delete this subDirectory. The
problem it solves is that if client again created the sub directory
after deleting, then we will end up applying that op without this
change.

**4.) Enabled the Fuzz tests which I wrote earlier, and they work now
:)**