feat(container-runtime): Prefer entryPoint for Summarizer (#14681)
## Description

(Taking over #14621 while Tony is OOF to get it into the next major
release).

Makes the default entryPoint for the summarizer clients the summarizer.
This reduces the coupling of the summarizer on the request pattern. Once
this change saturates, we can entirely drop the request pattern for the
summarizer.

Also removes some long-deprecated code, and an unused and non-functional
class(SummarizerHandle)

And moves the 2.0.0-internal.3.4.0 section in `BREAKING.md` below the
2.0.0-internal.4.0.0 section.

## Breaking Changes

- Removes deprecated interface `IProvideSummarizer`.
- Removes deprecated const `ISummarizer` (**not** the `ISummarizer`
interface).
- Makes it so `ISummarizer` doesn't extend `IFluidLoadable` nor
`Partial<IProvideSummarizer>` anymore.

## Reviewer guidance

Trying to get this in before we release 2.0.0-internal.4.0.0 (i.e. today
😄 before the release tomorrow).

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>