R11s/summary aliasing (#14401)
#13100 was getting out of hand with a refactor to Scenario Runner with
validation of this change. Instead, splitting out the scenario runner
testing portion of this work to a later time.

## Description

Currently, Scribe accepts 2 possible parent summaries: 1) the last
client summary, or 2) the latest service summary. However, there could
be multiple service summaries generated between the last client summary
and the latest service summary, and they are all valid parents.

The client runtime does not know about service summaries via SummaryAcks
(because service summaries do not generate SummaryAck ops), so service
summaries can be written without the knowledge of the runtime. This can
cause the runtime to propose an "outdated" service summary as a parent
summary. Currently, this is NACKed, forcing an inefficient recovery
process. Instead, we should allow the outdated service summary to be a
valid parent summary.

## Reviewer Guidance

The goal here is to track service summaries that have been written
between latest client summary and latest service summary within the
Scribe checkpoint data (persisted in DB and in summaries'
serviceProtocol blob). For example, if client summary A is written,
followed by service summaries B, C, D, then we will store the service
summary handles for B, C, and D and allow them to be referenced as
parent summaries for the next client summary.

### Prettier formatting

Looks like prettier formatting decided to change all the spaces to tabs
in SummaryWriter. I will drop a comment where there are actual changes