Fixed issue where if a deleted blob is de-duped, GC would not know about it (#14391)
## Bug
Blob de-duping has the following issues w.r.t. GC:
- In tombstone mode, after a blob has been tombstoned, its localId1 and
storageId1 are both added to the tombstone list. If this blob is
de-duped, a reference from localId2 will be added to storageId1. GC will
be notified of this and it will log error / throw because it thinks a
tombstoned blob was referenced.
- In sweep mode, the same problem as the above exists. The difference is
that the blob is added to a deleted list and an error is thrown because
GC thinks a deleted blob was referenced. In addition, the deleted
storageId1 will not be removed from the deleted list ever since GC only
ever appends to the list. This means that even after the new reference
from localId2 -> storageId1 is added, the blob will still be in the
deleted list. This should not have any functionality impact because a
blob will never be referenced by its storage id. However, its still
incorrect and can lead to other issues.

## Fix
Today, BlobManager gives both local and storage ids to GC when
getGCData() is called and they both participate in GC. However, with
[this change](https://github.com/microsoft/FluidFramework/pull/13486), a
storage id is never exposed by the blob manager, and so, it can never
make a blob referenced / unreferenced. In other words, it does not need
to participate in GC.

The fix is for the BlobManager to only return local ids in getGCData().
Storage ids then will never be tombstoned / deleted and the bug
described above will not happen.
BlobManager does an additional step in the sweep process - After local
blob ids have been deleted (since they participate in GC), for the
corresponding storage ids, it checks whether any localId -> storageId
entry exists. If there isn't an entry, the storage id is deleted as
well.
Another side benefit of this change is that BlobManager does not need to
notify GC of localId -> storageId references. In other words, GC does
not have to deal with blob de-duping at all.

[AB#2746](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2746)

## Alternative fix which was not done
When a reference from localId -> storageId is added, GC detects that and
doesn't log any tombstone or sweep errors. Additionally, it removes the
storageId from the deleted list. This solution is simple but it feels a
little hacky because GC now has to understand blob manager specifics.