ci: Fix version comparison in release step (#14701)
Cherry pick of #14692 to main.

We don't fully understand why, but the node 16/npm upgrade broke the old
way of comparing versions. This check is only done as a safeguard to
make sure that packages we're about to publish all have a version
matching the version we're releasing. The check is only done when
releasing.

I converted the script that checks the current version to a CJS script
to help with shell escaping issues. This change worked in manual
testing.