ci: Run prettier in repo-policy-check pipeline (#14703)
There's a gap in our prettier coverage in CI. If you edit a file in the
root, or, say, the `scripts` folder, it won't trigger a CI job that
includes prettier. The reason is that the client release group pipeline
is responsible for running prettier on the root files, but it's not
triggered by edits outside the client release group/pipeline.

There is a CI pipeline that runs on every PR: repo-policy-check. I fixed
this by adding a step to that pipeline that runs `npm run
prettier:root`. I also added pnpm store caching to the repo-policy-check
job and removed the layer-check and policy-check from the client
pipeline since those checks are repo-wide and done already in
repo-policy-check.