ci: Include build-tools READMEs in extraneous files check (#14673)
The build-tools readmes were excluded from the check originally because
they were modified during a build because the oclif readme generator
included the version, which differs every build. That is no longer the
case because we use a custom readme generation process now.

Despite this, they were still excluded, so this change corrects that
mistake.