Container docs rewrite and reorganization (#14527)
## Description

A rewrite and reorganization of the basic container docs:

1. The old article has been split in two. Most of the container state
and container event content has been moved to a new article.
2. The material in the new article has been greatly expanded. State
diagrams are included.
3. The new article will appear immediately below the old one in the ToC.
4. As explained in the articles, I replaced the concept of "attach-"
with "publish-", except when mentioning actual APIs.
5. I found that there was no need to speak of "loading" distinct from
"connecting". (In this respect, note that the article makes clear that
the **connected** state is not reached unless the **CatchingUp** phase
has completed, so "connected" implies caught up and, thus, loaded.) If
that decision is endorsed by reviewers, we should have another PR that
replaces most "loading" talk in other articles with "connecting" talk.
6. I did NOT include any information about what happens if there's an
error in, say, the **attaching** or **EstablishingConnection** states.
Including this would greatly complicate the story (and the state
diagrams) and harm the article's ability to provide an understanding the
states. If reviewers want to add this kind of content, I can add it to
another draft, but it should probably be separated into its own section,
with perhaps more complex versions of the state diagrams.

---------

Co-authored-by: Tyler Butler <tylerbu@microsoft.com>
Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>