fix(bump): Correctly save package.json of bumped packages (#14727)
I accidentally broke the package.json save logic in #14481. I used the
updatePackageJson function incorrectly. In this case,
`Package.savePackageJson`, we want to save the in-memory data to the
file, not read the file, process it, then write it back.

To fix this, I factored out functions to read JSON and return both
PackageJson objects and the detected indentation to reuse the common
logic. I also factored out the logic to sort and write the file. Then I
updated the Package class constructor to call the new function and
stores the indentation of the source JSON in a private property to use
when writing the file.

I also updated `updatePackageJsonFile` to use the new functions, so all
the common logic is shared.