build: Extract markdown-magic library into a repo-private package (#14629)
Does not change when / where the tooling is invoked, it simply hoists
the underlying logic into its own tooling library.

Subsequent changes will update the patterns for when / where we invoke
the tooling to better encapsulate doc generation in different parts of
the repo.

Notes:
* Abstracting the logic required updating how `INCLUDE` pragma paths are
represented. Previously, these were all written relative to the repo
root. Now they are relative to the document containing the pragma.
* Since the home of the logic changes, the generated content notices
added to documents was updated.
* A few pragma names were changed to make the usage semantics a bit more
explicit.
* The default value of `usesTinylious` for "Getting Started" sections
was changed from `false` to `true`, as the majority of our uses specify
true.

All of the changes to documentation files should be due to one of the
above - no other functional / content changes are expected.

---------

Co-authored-by: Tyler Butler <tyler@tylerbutler.com>
Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>