ci: Test for invalid Fluid internal version ranges (#14751)
Related to #14727.

I recently wrote a bug in the build tools that did not bump packages
properly in the pipeline. In particular, fluid internal versions used
dependency ranges with a `^` which won't work with that version scheme.

This change adds a check to the pipeline after we set versions to check
if any are using `^`.